[skip ci] Refactor CancellationPolicy
